DL Pipeline for 3D Muscle Segmentation/Final app/_requirements.txt
Libraries:
------------------------------
os
time
tqdm
numpy
torch
pandas
torchio
nibabel
scipy.ndimage


